+++ app/Http/Controllers/UserController.php
@@ -6,6 +6,8 @@ use App\Http\Requests\StoreUserRequest;
 use App\Http\Requests\UpdateUserRequest;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
@@ -15,6 +17,10 @@ class UserController extends Controller
      */
     public function index(Request $request)
     {
+        if (Gate::forUser(Auth::user())->denies('user-view')) {
+            return abort(403);
+        }
+
         $users = User::when($request->has('q'), function($query) use ($request){
             $query->where(function ($builder) use ($request) {
                 $builder->where('nome', 'like', '%' . $request->q . '%')
@@ -32,6 +38,10 @@ class UserController extends Controller
      */
     public function create()
     {
+        if (Gate::forUser(Auth::user())->denies('user-create')) {
+            return abort(403);
+        }
+
         $user = null;
         $disabled = false;
 
@@ -43,6 +53,10 @@ class UserController extends Controller
      */
     public function store(StoreUserRequest $request)
     {
+        if (Gate::forUser(Auth::user())->denies('user-create')) {
+            return abort(403);
+        }
+
         $fields = $request->only([
             'name', 'email'
         ]);
@@ -59,6 +73,10 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
+        if (Gate::forUser(Auth::user())->denies('user-view')) {
+            return abort(403);
+        }
+
         $disabled = true;
 
         return view('user.form', ['user' => $user, 'disabled' => $disabled]);
@@ -69,6 +87,10 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        if (Gate::forUser(Auth::user())->denies('user-edit')) {
+            return abort(403);
+        }
+
         $disabled = false;
 
         return view('user.form', ['user' => $user, 'disabled' => $disabled]);
@@ -79,6 +101,10 @@ class UserController extends Controller
      */
     public function update(UpdateUserRequest $request, User $user)
     {
+        if (Gate::forUser(Auth::user())->denies('user-edit')) {
+            return abort(403);
+        }
+
         $fields = $request->only([
             'name', 'email'
         ]);
@@ -97,6 +123,10 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        if (Gate::forUser(Auth::user())->denies('user-delete')) {
+            return abort(403);
+        }
+
         $user->delete();
 
         return redirect()->route('user.index')->with('flash_success', 'Usuário excluído com sucesso!');
